Module oak_proto_rust::oak

Modules

attestation

containers

crypto

debug

functions

key_provisioning

session

Structs

HexDigest

Similar to RawDigest, but contains hex-encoded hashes for the sake of better readability and copy-pastability. The set of all possible RawDigest and the set of all possible HexDigests are in a bijective correspondence, by just hex-encoding or hex-decoding each field separately.

RawDigest

Contains various digest formats for the same underlying file. Mixing digests from various files is “undefined behavior”. There is no distinction between empty and not set, it means the same.