Module oak_proto_rust::oak::attestation::v1

Modules

attestation_results

Nested message and enum types in AttestationResults.

binary_reference_value

Nested message and enum types in BinaryReferenceValue.

endorsement

Nested message and enum types in Endorsement.

endorsements

Nested message and enum types in Endorsements.

expected_digests

Nested message and enum types in ExpectedDigests.

expected_values

Nested message and enum types in ExpectedValues.

extracted_evidence

Nested message and enum types in ExtractedEvidence.

kernel_binary_reference_value

Nested message and enum types in KernelBinaryReferenceValue.

reference_values

Nested message and enum types in ReferenceValues.

root_layer_data

Nested message and enum types in RootLayerData.

text_expected_value

Nested message and enum types in TextExpectedValue.

text_reference_value

Nested message and enum types in TextReferenceValue.

verifying_key_reference_value

Nested message and enum types in VerifyingKeyReferenceValue.

Structs

AmdAttestationReport

Values extracted from an AMD SEV-SNP attestation report.

AmdSevExpectedValues

AmdSevReferenceValues

AmdSevSnpEndorsement

ApplicationEndorsement

ApplicationKeys

Keys used by the application to derive encryption session keys and to sign arbitrary data. Each of the certificates contains the final layer’s measurement as additional claims.

ApplicationLayerData

Values extracted from the evidence that represents an application running under the Oak Restricted Kernel.

ApplicationLayerEndorsements

ApplicationLayerExpectedValues

The expected binary and configuration digests for an application layer.

ApplicationLayerReferenceValues

Represents an application running under Oak Restricted Kernel.

AttestationResults

Represents a verification result. Can be extended to return certain measurements and other detail to the client for further processing. Nomenclature follows RFC 9334.

BinaryReferenceValue

CbData

Reserved for future use.

CbEndorsements

CbExpectedValues

CbReferenceValues

CertificateAuthority

Message for passing embedded certificate authority information between layers. Will never appear in the evidence that is sent to the client.

ClaimReferenceValue

Specifies a list of claim types. Claims are assertions about artifacts made by the endorsing entity. An overview of the claims maintained by Oak can be found at: https://github.com/project-oak/oak/tree/main/docs/tr/claim

ContainerEndorsement

ContainerLayerData

Values extracted from the evidence that represents the Container Runtime Bundle used in Oak Containers.

ContainerLayerEndorsements

ContainerLayerExpectedValues

The expected bundle and configuration digests for a container layer.

ContainerLayerReferenceValues

DiceData

Message that is sent between DICE layers.

Digests

Verifies that the field contains at least one of the given digests. No checks are performed if this is empty. A match in at least one digest is considered a success.

Endorsement

EndorsementDetails

Details about the endorsement statement which can be passed across FFI boundaries.

EndorsementReferenceValue

Verifies the transparency log entry, including signatures and the digest.

Endorsements

This message contains statements that some entity (e.g. a hardware provider) vouches for the integrity of claims about the TEE or the software running on it.

Event

Represents an event intended for inclusion in attestation. For example, in an attested measured boot, each event is a reference to the code identity of the boot layer being launched next. An Event message contain what’s necessary for an attestation verifier to verify the Event against a Reference Value. TODO: b/333748757 - Make other CB layers use this definition.

EventAttestationResults

Attestation verification results for an individual event. TODO: b/366419879 - Implement descriptive per-event attestation results.

EventData

Values extracted from the evidence that represents an event.

EventExpectedValues

Represents digest of an event.

EventLog

A sequence of Events intended for inclusion in attestation evidence.

EventReferenceValues

Represents digest of an event.

Evidence

Attestation Evidence used by the client to the identity of firmware and software running inside a Trusted Execution Environment.

ExpectedDigests

ExpectedRegex

ExpectedStringLiterals

ExpectedValues

ExtractedEvidence

Evidence values extracted from attestation evidence during verification.

FakeAttestationReport

Values extracted from a fake attestation report when not running in a TEE.

FileReferenceValue

Reference value for a file including its digests.

FirmwareAttachment

The Transparent Release attachment for Oak Stage 0. Measurements are produced with: https://github.com/project-oak/oak/tree/main/snp_measurement

FirmwareEndorsement

InsecureExpectedValues

InsecureReferenceValues

IntelTdxAttestationReport

Values extracted from an Intel TDX attestation report.

IntelTdxExpectedValues

IntelTdxReferenceValues

KernelAttachment

The Transparent Release attachment for Oak Containers Linux kernel. Measurements are produced with: https://github.com/project-oak/oak/tree/main/oak_kernel_measurement

KernelBinaryReferenceValue

Follows the lines of BinaryReferenceValue, but provides a custom proto to facilitate the digest matching.

KernelDigests

Similar to the Digests message, but allows to specify digests for the split components of the bzImage separately.

KernelEndorsement

KernelExpectedValues

The expected values for kernel image and setup data, computed from previously provided endorsements and reference values.

KernelLayerData

Values extracted from the the kernel layer evidence, as measured by stage0.

KernelLayerEndorsements

KernelLayerExpectedValues

Reference values of the kernel layer, as measured by stage0.

KernelLayerReferenceValues

Reference values of the kernel layer, as measured by stage0.

LayerEvidence

DICE layer evidence containing a certificate signed by the previous layer.

OakContainersData

Values extracted from the evidence for an Oak Containers instance.

OakContainersEndorsements

OakContainersExpectedValues

OakContainersReferenceValues

OakRestrictedKernelData

Values extracted from the evidence for a restricted kernel application.

OakRestrictedKernelEndorsements

OakRestrictedKernelExpectedValues

OakRestrictedKernelReferenceValues

OakStandaloneData

Oak Standalone currently skips all attestation

OrchestratorMeasurements

All the related measurements for Oak Container’s Stage 1.

RawDigests

ReferenceValues

Regex

RootLayerData

Values extracted from the root layer evidence.

RootLayerEndorsements

RootLayerEvidence

Evidence generated by the Layer0.

RootLayerExpectedValues

RootLayerReferenceValues

Signature

SignedEndorsement

A signed endorsement which is optionally published.

SkipVerification

Stage0Measurements

All the related measurements for Stage 0.

Stage1Measurements

All the related measurements for Oak Container’s Stage 1.

StringLiterals

A match in at least one value is considered a success. At least one value must be specified, otherwise verification fails.

SystemEndorsement

SystemLayerData

Values extracted from the evidence that represents the Oak Containers system image.

SystemLayerEndorsements

SystemLayerExpectedValues

The expected binary digests for a system layer image.

SystemLayerReferenceValues

TcbVersion

The versions of the components in the AMD SEV-SNP platform Trusted Compute Base (TCB).

TextExpectedValue

TextReferenceValue

Reference value to match text via endorsement, or directly via constants or a regular expression.

TransparentReleaseEndorsement

A Transparent Release endorsement for a binary which includes the actual endorsement, a signature over it, and optionally a transparency log entry. Don’t use this message in new code, use SignedEndorsement instead. SignedEndorsement supersedes TransparentReleaseEndorsement.

Validity

Specifies a temporal range of validity for an endorsement.

VerificationSkipped

VerifyingKey

VerifyingKeyReferenceValue

Reference values that control how the endorsement is verified.

VerifyingKeySet

Set of keys currently needed for verification. Will contain one element most of the time, but there may be more during key rotation/revocation. To revoke a key, simply add a new one, allow a grace period, and then don’t pass the old key here.

Enums

KeyType

TeePlatform

This proto defines the layered DICE Attestation Evidence.