Struct oak_proto_rust::oak::functions::config::WasmtimeConfig

pub struct WasmtimeConfig {
    pub pooling_strategy: Option<PoolingAllocationConfig>,
    pub static_memory_maximum_size: Option<u64>,
    pub static_memory_guard_size: Option<u64>,
    pub dynamic_memory_guard_size: Option<u64>,
    pub dynamic_memory_reserved_for_growth: Option<u64>,
    pub memory_init_cow: Option<bool>,
}

Settings specific to the Wasmtime engine.

These fields are explicitly marked as optional so that we could detect their presence (as sometimes you do want to specify zero as a value). If a field is left unspecified, it will leave whatever Wasmtime has as the default.

See https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html for more details.

Fields
pooling_strategy: Option<PoolingAllocationConfig>

If specified, switches to a pooling allocation strategy. If not specified, leaves as default (on demand). https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.allocation_strategy

static_memory_maximum_size: Option<u64>

https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.static_memory_maximum_size

static_memory_guard_size: Option<u64>

https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.static_memory_guard_size

dynamic_memory_guard_size: Option<u64>

https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.dynamic_memory_guard_size

dynamic_memory_reserved_for_growth: Option<u64>

https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.dynamic_memory_reserved_for_growth

memory_init_cow: Option<bool>

https://docs.rs/wasmtime/latest/wasmtime/struct.Config.html#method.memory_init_cow

Implementations

impl WasmtimeConfig

pub fn static_memory_maximum_size(&self) -> u64

Returns the value of static_memory_maximum_size, or the default value if static_memory_maximum_size is unset.

pub fn static_memory_guard_size(&self) -> u64

Returns the value of static_memory_guard_size, or the default value if static_memory_guard_size is unset.

pub fn dynamic_memory_guard_size(&self) -> u64

Returns the value of dynamic_memory_guard_size, or the default value if dynamic_memory_guard_size is unset.

pub fn dynamic_memory_reserved_for_growth(&self) -> u64

Returns the value of dynamic_memory_reserved_for_growth, or the default value if dynamic_memory_reserved_for_growth is unset.

pub fn memory_init_cow(&self) -> bool

Returns the value of memory_init_cow, or the default value if memory_init_cow is unset.