Module oak_proto_rust::oak::session::v1

Modules

handshake_request

Nested message and enum types in HandshakeRequest.

handshake_response

Nested message and enum types in HandshakeResponse.

request_wrapper

Nested message and enum types in RequestWrapper.

response_wrapper

Nested message and enum types in ResponseWrapper.

session_request

Nested message and enum types in SessionRequest.

session_response

Nested message and enum types in SessionResponse.

Structs

AttestRequest

Request message for the remote attestation.

AttestResponse

Response message for the remote attestation.

EncryptedMessage

Message for encrypted data exchange after a secure session is established.

EndorsedEvidence

Endorsed evidence contains an attestation evidence provided by the enclave and the corresponding attestation endorsements provided by the hostlib.

GetEndorsedEvidenceRequest

GetEndorsedEvidenceResponse

HandshakeRequest

Request message for the crypto handshake request needed to establish a set of session keys.

HandshakeResponse

Response message for the crypto handshake request needed to establish a set of session keys.

InvokeRequest

InvokeResponse

NoiseHandshakeMessage

Noise handshake message containing fields for all handshake patterns. <http://www.noiseprotocol.org/noise.html#handshake-patterns>

PlaintextMessage

Message with decrypted content (not to be transmitted over the wire).

RequestWrapper

ResponseWrapper

SessionBinding

Message that binds the Noise session (and optionally other data) to the Attestation Evidence.

SessionRequest

Request message for the Oak protocol attested secure session. This message is a wrapper containing different message types including: attestation, handshake and encrypted data exchange.

SessionRequestWithSessionId

Wrapper around SessionRequest that is used in cases where it is necessary to identify a contiguous session across multiple invocations/streams.

SessionResponse

Response message for the Oak protocol attested secure session. This message is a wrapper containing different message types including: attestation, handshake and encrypted data exchange.